Fix movement and check for wall collision
Before, the sprite was being drawn at an offset and rotated incorrectly
due to the rotation being set around the bottom-right corner instead of
the center of the sprite.
Wall collision is smart and checks for transparency; the transparent
parts of the sprite can go through black walls.

Also made Stage 3 simpler. More stages in increasing difficulty will
be added later.

+++ Game/src/me/sergey/sprites/GButton.java
@@ -10,7 +10,7 @@ public abstract class GButton extends Base implements EventHandler<MouseEvent>{
         this(gc, path, -1, -1);
     }
     
-    public GButton(GraphicsContext gc, String path, int x, int y){
+    public GButton(GraphicsContext gc, String path, double x, double y){
         super(gc, path, x, y);
         this.gc.getCanvas().getScene().addEventHandler(MouseEvent.MOUSE_CLICKED, this);
     }
